Finish Lab03.

+++ lab03/src/adventure/SpeciesListStage.java
@@ -98,13 +98,13 @@ public class SpeciesListStage implements AdventureStage {
      */
     public static double arraySimilarity(List<String> listOne, List<String> listTwo) {
         List<String> copy = new ArrayList<>(listOne);
-        int similarObjects = 0;
+        double similarObjects = 0.0;
         for (String o : listTwo) {
             if (copy.contains(o)) {
                 similarObjects++;
                 copy.remove(o);
             }
         }
-        return similarObjects / listOne.size();
+        return similarObjects / (double) listOne.size();
     }
 }
